requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
certifi==2022.6.15.2
cffi==1.15.1
channels==2.4.0
channels-redis==2.4.2
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.1
daphne==2.4.1
Django==4.1.1
django-braces==1.15.0
django-embed-video==1.4.5
django-filter==22.1
django-memcache-status==1.3
djangorestframework==3.13.1
hiredis==2.0.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.12.0
incremental==21.3.0
isort==5.10.1
Markdown==3.4.1
msgpack==0.6.2
Pillow==9.2.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pymemcache==3.5.2
pyOpenSSL==22.0.0
python-memcached==1.59
pytz==2022.2.1
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.8.0
txaio==22.2.1
typing-extensions==4.3.0
urllib3==1.26.12
uWSGI==2.0.18
zipp==3.8.1
zope.interface==5.4.0
